User profiles

+++ internal/models/user.go
@@ -17,8 +17,8 @@ type User struct {
 	DateOfBirth    time.Time          `json:"dateOfBirth" level:"1"`   // the user's date of birth, used for calculating age
 	ProfileFields  []UserProfileField `json:"profile"`                 // fields of the user's profile
 	// ZIPCode        string             `json:"zipCode" level:"1"`       // the user's zip code, used to find nearby opportunities
-	LocationLatitude  float32 `json:"-"` // the latitude of the user's city
-	LocationLongitude float32 `json:"-"` // the longitude of the user's city
+	LocationLatitude  float64 `json:"-"` // the latitude of the user's city
+	LocationLongitude float64 `json:"-"` // the longitude of the user's city
 }
 
 // UserRepository represents a repository of users.
